requirements.txt
asgiref==3.7.2
boto3==1.28.36
botocore==1.31.36
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
Django==4.2.4
django-appconf==1.0.5
django-crispy-forms==2.0
django-imagefit==0.8.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
